added nepali-date as dependency for handling dates and moved types to seperate file
