--- src/utils/data.ts
@@ -1,247 +1,0 @@
-export const startDateAD = 19430414;
-export const startDateBS = 20000101;
-
-export const lowerADDate = -62135596800000;
-
-export const yearMonthDays = [
-  [30, 32, 31, 32, 31, 30, 30, 30, 29, 30, 29, 31], // 2000 BS
-  [31, 31, 32, 31, 31, 31, 30, 29, 30, 29, 30, 30],
-  [31, 31, 32, 32, 31, 30, 30, 29, 30, 29, 30, 30],
-  [31, 32, 31, 32, 31, 30, 30, 30, 29, 29, 30, 31],
-  [30, 32, 31, 32, 31, 30, 30, 30, 29, 30, 29, 31],
-  [31, 31, 32, 31, 31, 31, 30, 29, 30, 29, 30, 30],
-  [31, 31, 32, 32, 31, 30, 30, 29, 30, 29, 30, 30],
-  [31, 32, 31, 32, 31, 30, 30, 30, 29, 29, 30, 31],
-  [31, 31, 31, 32, 31, 31, 29, 30, 30, 29, 29, 31],
-  [31, 31, 32, 31, 31, 31, 30, 29, 30, 29, 30, 30],
-  [31, 31, 32, 32, 31, 30, 30, 29, 30, 29, 30, 30], // 2010
-  [31, 32, 31, 32, 31, 30, 30, 30, 29, 29, 30, 31],
-  [31, 31, 31, 32, 31, 31, 29, 30, 30, 29, 30, 30],
-  [31, 31, 32, 31, 31, 31, 30, 29, 30, 29, 30, 30],
-  [31, 31, 32, 32, 31, 30, 30, 29, 30, 29, 30, 30],
-  [31, 32, 31, 32, 31, 30, 30, 30, 29, 29, 30, 31],
-  [31, 31, 31, 32, 31, 31, 29, 30, 30, 29, 30, 30],
-  [31, 31, 32, 31, 31, 31, 30, 29, 30, 29, 30, 30],
-  [31, 32, 31, 32, 31, 30, 30, 29, 30, 29, 30, 30],
-  [31, 32, 31, 32, 31, 30, 30, 30, 29, 30, 29, 31],
-  [31, 31, 31, 32, 31, 31, 30, 29, 30, 29, 30, 30], // 2020
-  [31, 31, 32, 31, 31, 31, 30, 29, 30, 29, 30, 30],
-  [31, 32, 31, 32, 31, 30, 30, 30, 29, 29, 30, 30],
-  [31, 32, 31, 32, 31, 30, 30, 30, 29, 30, 29, 31],
-  [31, 31, 31, 32, 31, 31, 30, 29, 30, 29, 30, 30],
-  [31, 31, 32, 31, 31, 31, 30, 29, 30, 29, 30, 30],
-  [31, 32, 31, 32, 31, 30, 30, 30, 29, 29, 30, 31],
-  [30, 32, 31, 32, 31, 30, 30, 30, 29, 30, 29, 31],
-  [31, 31, 32, 31, 31, 31, 30, 29, 30, 29, 30, 30],
-  [31, 31, 32, 31, 32, 30, 30, 29, 30, 29, 30, 30],
-  [31, 32, 31, 32, 31, 30, 30, 30, 29, 29, 30, 31], // 2030
-  [30, 32, 31, 32, 31, 30, 30, 30, 29, 30, 29, 31],
-  [31, 31, 32, 31, 31, 31, 30, 29, 30, 29, 30, 30],
-  [31, 31, 32, 32, 31, 30, 30, 29, 30, 29, 30, 30],
-  [31, 32, 31, 32, 31, 30, 30, 30, 29, 29, 30, 31],
-  [30, 32, 31, 32, 31, 31, 29, 30, 30, 29, 29, 31],
-  [31, 31, 32, 31, 31, 31, 30, 29, 30, 29, 30, 30],
-  [31, 31, 32, 32, 31, 30, 30, 29, 30, 29, 30, 30],
-  [31, 32, 31, 32, 31, 30, 30, 30, 29, 29, 30, 31],
-  [31, 31, 31, 32, 31, 31, 29, 30, 30, 29, 30, 30],
-  [31, 31, 32, 31, 31, 31, 30, 29, 30, 29, 30, 30], // 2040
-  [31, 31, 32, 32, 31, 30, 30, 29, 30, 29, 30, 30],
-  [31, 32, 31, 32, 31, 30, 30, 30, 29, 29, 30, 31],
-  [31, 31, 31, 32, 31, 31, 29, 30, 30, 29, 30, 30],
-  [31, 31, 32, 31, 31, 31, 30, 29, 30, 29, 30, 30],
-  [31, 32, 31, 32, 31, 30, 30, 29, 30, 29, 30, 30],
-  [31, 32, 31, 32, 31, 30, 30, 30, 29, 29, 30, 31],
-  [31, 31, 31, 32, 31, 31, 30, 29, 30, 29, 30, 30],
-  [31, 31, 32, 31, 31, 31, 30, 29, 30, 29, 30, 30],
-  [31, 32, 31, 32, 31, 30, 30, 30, 29, 29, 30, 30],
-  [31, 32, 31, 32, 31, 30, 30, 30, 29, 30, 29, 31], // 2050
-  [31, 31, 31, 32, 31, 31, 30, 29, 30, 29, 30, 30],
-  [31, 31, 32, 31, 31, 31, 30, 29, 30, 29, 30, 30],
-  [31, 32, 31, 32, 31, 30, 30, 30, 29, 29, 30, 30],
-  [31, 32, 31, 32, 31, 30, 30, 30, 29, 30, 29, 31],
-  [31, 31, 32, 31, 31, 31, 30, 29, 30, 29, 30, 30],
-  [31, 31, 32, 31, 32, 30, 30, 29, 30, 29, 30, 30],
-  [31, 32, 31, 32, 31, 30, 30, 30, 29, 29, 30, 31],
-  [30, 32, 31, 32, 31, 30, 30, 30, 29, 30, 29, 31],
-  [31, 31, 32, 31, 31, 31, 30, 29, 30, 29, 30, 30],
-  [31, 31, 32, 32, 31, 30, 30, 29, 30, 29, 30, 30], // 2060
-  [31, 32, 31, 32, 31, 30, 30, 30, 29, 29, 30, 31],
-  [30, 32, 31, 32, 31, 31, 29, 30, 29, 30, 29, 31],
-  [31, 31, 32, 31, 31, 31, 30, 29, 30, 29, 30, 30],
-  [31, 31, 32, 32, 31, 30, 30, 29, 30, 29, 30, 30],
-  [31, 32, 31, 32, 31, 30, 30, 30, 29, 29, 30, 31],
-  [31, 31, 31, 32, 31, 31, 29, 30, 30, 29, 29, 31],
-  [31, 31, 32, 31, 31, 31, 30, 29, 30, 29, 30, 30],
-  [31, 31, 32, 32, 31, 30, 30, 29, 30, 29, 30, 30],
-  [31, 32, 31, 32, 31, 30, 30, 30, 29, 29, 30, 31],
-  [31, 31, 31, 32, 31, 31, 29, 30, 30, 29, 30, 30], // 2070
-  [31, 31, 32, 31, 31, 31, 30, 29, 30, 29, 30, 30],
-  [31, 32, 31, 32, 31, 30, 30, 29, 30, 29, 30, 30],
-  [31, 32, 31, 32, 31, 30, 30, 30, 29, 29, 30, 31],
-  [31, 31, 31, 32, 31, 31, 30, 29, 30, 29, 30, 30],
-  [31, 31, 32, 31, 31, 31, 30, 29, 30, 29, 30, 30],
-  [31, 32, 31, 32, 31, 30, 30, 30, 29, 29, 30, 30],
-  [31, 32, 31, 32, 31, 30, 30, 30, 29, 30, 29, 31],
-  [31, 31, 31, 32, 31, 31, 30, 29, 30, 29, 30, 30],
-  [31, 31, 32, 31, 31, 31, 30, 29, 30, 29, 30, 30],
-  [31, 32, 31, 32, 31, 30, 30, 30, 29, 29, 30, 30], // 2080
-  [31, 32, 31, 32, 31, 30, 30, 30, 29, 30, 29, 31], // 2081 BS
-  [30, 32, 31, 32, 31, 30, 30, 30, 29, 30, 30, 30],
-  [31, 31, 32, 31, 31, 31, 30, 29, 30, 29, 30, 30],
-  [31, 31, 32, 31, 31, 30, 30, 30, 29, 30, 30, 30],
-  [31, 32, 31, 32, 30, 31, 30, 30, 29, 30, 30, 30],
-  [30, 32, 31, 32, 31, 30, 30, 30, 29, 30, 30, 30],
-  [31, 31, 32, 31, 31, 31, 30, 30, 29, 30, 30, 30],
-  [30, 31, 32, 32, 30, 31, 30, 30, 29, 30, 30, 30],
-  [30, 32, 31, 32, 31, 30, 30, 30, 29, 30, 30, 30],
-];
-
-export const baishakOne = [
-  19430414, 19440413, 19450413, 19460413, 19470414, 19480413, 19490413,
-  19500413, 19510414, 19520413, 19530413, 19540413, 19550414, 19560413,
-  19570413, 19580413, 19590414, 19600413, 19610413, 19620413, 19630414,
-  19640413, 19650413, 19660413, 19670414, 19680413, 19690413, 19700414,
-  19710414, 19720413, 19730413, 19740414, 19750414, 19760413, 19770413,
-  19780414, 19790414, 19800413, 19810413, 19820414, 19830414, 19840413,
-  19850413, 19860414, 19870414, 19880413, 19890413, 19900414, 19910414,
-  19920413, 19930413, 19940414, 19950414, 19960413, 19970413, 19980414,
-  19990414, 20000413, 20010414, 20020414, 20030414, 20040413, 20050414,
-  20060414, 20070414, 20080413, 20090414, 20100414, 20110414, 20120413,
-  20130414, 20140414, 20150414, 20160413, 20170414, 20180414, 20190414,
-  20200413, 20210414, 20220414, 20230414, 20240413, 20250414, 20260414,
-  20270414, 20280413, 20290414, 20300414, 20310415, 20320414, 20330414,
-];
-
-export const ENGLISH_WEEK = ['Su', 'Mo', 'Tu', 'We', 'Th', 'Fr', 'Sa'];
-export const ENGLISH_WEEK_FULL = [
-  'Sunday',
-  'Monday',
-  'Tuesday',
-  'Wednesday',
-  'Thursday',
-  'Friday',
-  'Saturday',
-];
-
-export const NEPALI_DIGITS = {
-  '0': '०',
-  '1': '१',
-  '2': '२',
-  '3': '३',
-  '4': '४',
-  '5': '५',
-  '6': '६',
-  '7': '७',
-  '8': '८',
-  '9': '९',
-};
-
-export const NEPALI_DIGITS_TO_ENG = {
-  '०': '0',
-  '१': '1',
-  '२': '2',
-  '३': '3',
-  '४': '4',
-  '५': '5',
-  '६': '6',
-  '७': '7',
-  '८': '8',
-  '९': '9',
-};
-
-export const NEPALI_WEEK = [
-  'आइत',
-  'सोम',
-  'मंगल',
-  'बुध',
-  'बिहि',
-  'शुक्र',
-  'शनि',
-];
-
-export const NEPALI_WEEK_FULL = [
-  'आइतबार',
-  'सोमबार',
-  'मंगलबार',
-  'बुधबार',
-  'बिहीबार',
-  'शुक्रबार',
-  'शनिबार',
-];
-
-export const ENGLISH_NEPALI_MONTH = [
-  'Baisakh',
-  'Jestha',
-  'Asar',
-  'Shrawan',
-  'Bhadra',
-  'Aswin',
-  'Kartik',
-  'Mangsir',
-  'Poush',
-  'Magh',
-  'Falgun',
-  'Chaitra',
-];
-
-export const NEPALI_MONTH = [
-  'बैशाख',
-  'जेठ',
-  'असार',
-  'श्रावण',
-  'भाद्र',
-  'आश्विन',
-  'कार्तिक',
-  'मंसिर',
-  'पौष',
-  'माघ',
-  'फाल्गुण',
-  'चैत्र',
-];
-
-export const AD_MONTH = [
-  'Jan',
-  'Feb',
-  'Mar',
-  'Apr',
-  'May',
-  'Jun',
-  'Jul',
-  'Aug',
-  'Sep',
-  'Oct',
-  'Nov',
-  'Dec',
-];
-
-export const AD_MONTH_FULL = [
-  'January',
-  'February',
-  'March',
-  'April',
-  'May',
-  'June',
-  'July',
-  'August',
-  'September',
-  'October',
-  'November',
-  'December',
-];
-
-export const AD_MONTH_NEPALI = [
-  'जनवरी',
-  'फेब्रुअरी',
-  'मार्च',
-  'अप्रिल',
-  'मे',
-  'जुन',
-  'जुलाई',
-  'अगस्ट',
-  'सेप्टेम्बर',
-  'अक्टोबर',
-  'नोभेम्बर',
-  'डिसेम्बर',
-];
-
-export const NEPALI_TODAY = 'आज';
